pyproject.toml
[tool.poetry]
name = "queremossaber"
version = "0.1.0"
description = ""
authors = ["Vitor Baptista <vitor@vitorbaptista.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
django = ">=2.1.6,<3"
django-widget-tweaks = "1.4.3"
django-environ = "0.4.5"
django-storages = "1.7.1"
boto3 = "1.9.143"
raven = "6.10.0"
psycopg2-binary = "<2.9"
libsass = "0.18.0"
whitenoise = "5.3.0"
gunicorn = "19.9.0"

[tool.poetry.dev-dependencies]
django-livesync = "0.5"
django-debug-toolbar = "1.11"
honcho = "1.0.1"
watchdog = "2.2.1"

[tool.poetry.group.dev.dependencies]
pytest-django = "^4.9.0"
pre-commit = "^3.8.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "web.settings"

requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
